pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kensa-ai"
version = "0.1.0"
description = "検査-AI: Open-source AI red teaming toolkit with ISO/IEC 42001 alignment"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Jim Lu", email = "jimlu@example.com"}
]
maintainers = [
    {name = "Jim Lu", email = "jimlu@example.com"}
]
keywords = [
    "ai",
    "security",
    "red-team",
    "llm",
    "testing",
    "adversarial",
    "iso-42001",
    "prompt-injection",
    "jailbreak",
    "ai-safety",
    "machine-learning",
    "cybersecurity"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Software Development :: Testing",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]
dependencies = [
    "aiohttp>=3.9.0",
    "click>=8.1.0",
    "httpx>=0.25.0",
    "jinja2>=3.1.0",
    "openai>=1.10.0",
    "anthropic>=0.18.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "tenacity>=8.2.0",
    "tiktoken>=0.5.0",
    "structlog>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "black>=23.12.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]
ml = [
    "transformers>=4.36.0",
    "torch>=2.1.0",
    "sentence-transformers>=2.2.0",
]

[project.scripts]
kensa = "kensa_ai.cli:main"
kensa-ai = "kensa_ai.cli:main"

[project.urls]
Homepage = "https://github.com/jimlu/kensa-ai"
Documentation = "https://github.com/jimlu/kensa-ai#readme"
Repository = "https://github.com/jimlu/kensa-ai"
Issues = "https://github.com/jimlu/kensa-ai/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.yml", "*.json", "*.html", "*.css"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.mypy_cache
    | \.ruff_cache
    | \.venv
    | build
    | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "S",   # flake8-bandit (security)
]
ignore = [
    "E501",  # line too long (handled by black)
    "S101",  # use of assert (OK in tests)
]

[tool.ruff.per-file-ignores]
"tests/*" = ["S101"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=kensa_ai --cov-report=term-missing"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["src/kensa_ai"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
